message done

+++ test/Datingapp/Client/src/app/messages/messages.component.ts
@@ -6,6 +6,7 @@ import { PaginationModule } from 'ngx-bootstrap/pagination';
 import { TimeagoModule } from 'ngx-timeago';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -23,7 +24,7 @@ pageNumber=1;
 pageSize=5;
 loading=false;
 
-constructor(private messageService: MessageService){
+constructor(private messageService: MessageService,private confirmService:ConfirmService){
   this.loadMessages();
 }
 
@@ -37,8 +38,13 @@ loadMessages(){
 }
 
 deleteMessage(id:number){
-  this.messageService.deleteMessage(id).subscribe(()=>{
-    this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+  this.confirmService.confirm('Confirm delete Message', 'This cannot be undone').subscribe(result=>{
+    if(result){
+
+      this.messageService.deleteMessage(id).subscribe(()=>{
+        this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+      })
+    }
   })
 }
